working, needs a fine tooth comb as cursor has killed itself,.
